Setup for the background merge thread
This required splitting the actual state (what we used to be struct BitRust)
into the state (BitRustState) and a wrapper (BitRust). The wrapper would
then spawn and reap any background threads while delegating database
calls to the internal BitRustState value.

What I do not like at this point is that there are two levels of locking
involved now. The BitRustState struct operations all acquire a RwLock on the
object before running. That should take care of concurrent accesses via
the database accessors (put/get/delete) of the BitRustState object.

However, in order to mutex the access to the state object itself, we
need a lock at the BitRust (wrapper) level. For example, the background
thread, after performing a merge, needs to atomically change the active
file of the underlying BitRustState.

At this point I'll let both locks be, but one of them can be gotten rid
of.