Push config down to BitRustState
BitRust was holding the Config object and constructing the BitRustState
with the data directory. But then I needed to get the max active file
size, and not having it in the BitRustState object makes it super
awkward. Besides, there is no good reason for the config to live in
BitRust (as opposed to in BitRustState).