Test for reading next BitRustDataRecord from a Read
Other tests were also subtly broken because of an unneeded trim() when
parsing guide file lines. The value of a record happens to be the last
field on a line, and isn't quoted. The trimming removes any trailing
spaces from the value part, making tests fail even if
BitRustDataRecord::next_from_file() is working fine.

At this point, I am not investing in changing the guide file format
(although that would be easy enough -- just have to make sure the
Python script encodes compatibly to any rust csv crate we choose to
use), but getting rid of the .trim() works well, since the guide file
has UNIX line endings with no blank line at the end of the file, and we
are splitting on '\n' anyway.