Move ActiveFile::tell() to its own impl
The default impl via ReadableFile::tell() uses self.file() to report the
current offset. self.file() for both ActiveFile and ReadableFile returns
the read handle. For ActiveFile, the write handle already is at the end,
so I am using it. I am not sure if this will imply a seek or not, maybe
it does, since it seems unreasonable for the implementation to keep track
of whether a file pointer is really at the end of a stream with multiple
writers to the same stream being possible.