synchronization: Remove 2-level locking from reads
The KeyDir structure has an rwlock internally, making its get/put/keys methods atomic.

For reads, once we have the entry from the keydir, we are going to read from an
immutable region of a data file (which the entry contains a pointer to). Another lock
is unnecessary since both writes and other concurrent reads can proceed once we have
the entry from the keydir. And getting the entry from the keydir is already synchronized
internally by the structure.

For writes, however, we must ensure that there is only one actually happening at a given
time, since otherwise concurrent writes can clobber the append only file being written
to.

So the BitRustState level mutex is needed to synchronize writes such that they don't
clobber each other, while the second, RWLock is also needed for writes to indicate when
they're changing the keydir, so that any concurrent reads will properly wait. In other
words, I don't think this bilevel locking can be done away with.