Merge: first cut
The merge process is responsible for compacting old data files into
(hopefully) a smaller set of data files. Here, we have an "inside out"
variant -- i.e., instead of walking the keys in the keydir and writing
records into a new set of merge files from there, we walk the old record
files in sorted order of their age, record by record, writing the new
merged files (and updating the keydir as we go).

The locking scheme is explained in the code, but apart from
brainstorming, I have not used a tool to proofcheck that it works.
Obviously it works today, tomorrow a fuzztest might prove otherwise :)

TODO
- Make the process consider if a previous merge actually yielded any
savings in disk space.
- Fuzz Test!