Implement a working merge algorithm
The previous algorithm that was based on file ids had a couple of flaws:

1. When merging a file that contains two occurrences of the same key
   (and supposing no later file contains this key) the older of the
   records would be reflected in the merge output. This is because after
   the first record is merged, the keydir would be updated to point to
   the merge output file, which has a higher id than the file being
   merged. This would make us skip the subsequent record with the same
   key, thinking that the keydir contains a "later" record.

2. Even if we could fix the above problem by making sure that all merge
   output file ids generated in the current merge phase do not count as
   more current than the merge input file ids, we would be momentarily
   putting stale data into the keydir.

The solution to this was to use record timestamps instead of file ids
for checking record age. This way, the keydir will be updated during
a merge *only* for the latest record, which is being pointed to by the
active keydir.

However, using second resolution timestamps as we were before is not
sufficient for this, and essentially leads to the same problem as above
for writes are made in the same second. System clock timestamps have
other problems any way, like potentially making the datadir not portable
across machines with substantial clock skew.

So we implement simple logical timestamps that are incremented for each
record. Across restarts, we can easily bring the clock back to the state
before shutdown by keeping track of the highest record id we read as we
do the initial scan to build the keydir.