Start RecordIO implementation
Based on a popular way of encoding delimited messages ("RecordIO"), this
impl encodes each record of type M as

  record_size        [8 bytes]
  record_size_crc32  [4 bytes]
  serialized_message [record_size bytes]
  message_crc32      [4 bytes]

therefore adding 16 bytes of overhead to each record. The sizes
themselves are not varint encoded.

These two traits, RecordReader and RecordAppender will replace the
ReadableFile impls and the inherent impl for ActiveFile, providing
a consistent recordio interface for both data and hint file records.

I had to put the tighter constraing of BufRead on the inner reader of
RecordReader so as to be able to signal EOF easily using the fill_buf()
method -- if it returns an empty buffer, EOF has been reached per
[docs][1]

[1]: https://doc.rust-lang.org/std/io/trait.BufRead.html#tymethod.fill_buf