Move ActiveFile and InactiveFile to Record{Appender,Reader} based recordio
This means that the migration to protobuf encoded messages is complete.

This also fixes a critical bug with merges, which would cause data loss
after merge. The scenario is this:

-- file id 0 --
foo  foo_0
foo  foo_1
-- file id 1 --
foo  foo_2
foo  foo_3
-- file id 2 -- [active file, empty]

Now if we start merge of files 0, 1, the merge output id is 3. In the
main merge loop, after writing a merge record, we used to check if the
current file id for that key in the keydir is the *same* as the current
file id being merged. The assumption was that if the currently being
merged file id is less than the keydir file id, we have a newer record
in the keydir. However, in the above scenario, the all legitimate merge
updates would be retreated because after writing the first merge record
(for value foo_0), the keydir file id becomes 3. But this does not mean
that the entry is latest!

So as a fix, we keep track of all merge file ids generated in a cycle
(currently only one, but later more once we start overflowing merge
files), and update the keydir if either the keydir entry file id is the
same as the file id of the currently in-merge file, or if the current
keydir entry points to a merge file from this cycle.