Fix bug with keydir loading
Instead of returning the updated offset for each record, we were
returning the *record size*, which was causing wrong entry offsets to be
loaded into the keydir. There was no test that exercised the restart
behaviour multiple times to catch this serious bug, that was likely
introduced during the protobuf migration.