Fix startup code to consider record timestamps to avoid time-travel
There was a bug in the startup code where we process files in increasing
order of ids, meaning that of multiple entries for the same key, the
last one in the highest file id will prevail.

This is wrong, since we can have lower file ids containing later
records. For example, say we have the following state:

  inactive file ids: [0, 1]
  active file id: 2

Now, without any writes, we call merge. This will create a merge files
starting at id 3, and files 0, 1 will be gone after the merge. Now
suppose we get a put for key "foo", which also exists in file id 3. This is
written to the active file, which has id 2. This means that there is a *later*
entry for "foo" in (2) than in (3). And in fact bitrust will reply with
the newer entry right after the put. However, if we close and reopen the
bitrust, the later put will be elided by the code which simply processes
records sequentially from files, in ascending order of file id.

The fix is easy: During startup, overwrite an existing keydir record
only if it is older than the candidate record.