Maintain writer offset instead of "seeking 0 bytes from end" everytime
I did not see significant performance improvement with this (benchmark
for putrandom_no_merge) was within error bounds. This is surprising as
I thought doing a seek on *every* put would be more expensive.

But then I realized I am on an SSD. I am not sure what exactly goes on
when we seek files stored on flash storage, but it is fair to say that
the cost of random access on flash is much much lower compared to that
on (spinning) disks. It would be interesting to actually run this
benchmark on disk.