fsync after each write
I wanted to first use the example_service to show that not fsync-ing can
cause data-loss. For instance, I would have a pilot program that sends
writes to a bitrust over RPC, keeping track of successful writes in
a local map. It would then randomly simulate a poweroff for the
container/VM running the server. The intention was to be able to reach
a state where data in the kernel disk cache was not written out even if
the corresponding write()s completed successfully.

To reliably simulate a cold power cut for this experiment, the
hypervisor (if running the server in a VM) must be told to drop
unwritten bytes from its cache. I am not even sure how to do this with
a docker container. Being able to do this experiment easily would have
made for a better commit message with a nice before/after story.

Whatever, we know that fsync()ing before acknowledging a write is the
correct default behaviour. Write perf does not seem affected in the
random puts bench:

getrandom_no_merge      time:   [1.9759 us 1.9952 us 2.0152 us]
                        change: [-1.0871% -0.2185% +0.7553%] (p = 0.65 > 0.05)
                        No change in performance detected.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

putrandom_no_merge      time:   [1.7528 us 2.1921 us 3.1138 us]
                        change: [-12.402% +20.944% +65.115%] (p = 0.26 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) high mild
  4 (4.00%) high severe