Do not generate protobuf code inside src
This is not recommended, and cargo publish fails unless you want to
proceed with --no-verify. See
https://github.com/rust-lang/cargo/issues/5073.

I am copying the approach taken in
https://github.com/Zondax/ledger-rs/commit/43908c4d033b8fe53b1cde76bf5c3db70b72fa7c,
where we generate the proto sources in OUT_DIR (an envvar available to
build scripts), and then communicating the list of generated mods to
rustc via a custom envvar.